+++ packages/dnb-ui-lib/src/components/global-status/__tests__/GlobalStatus.screenshot.test.js
@@ -9,23 +9,46 @@ import {
 } from '../../../core/jest/jestSetupScreenshots'
 
 describe('GlobalStatus screenshot', () => {
+  const style = { width: '20rem' }
   setupPageScreenshot({ url: '/uilib/components/global-status' })
-  it.skip('have to match the global-status with icon', async () => {
+  it('have to match the default state with custom content', async () => {
     const screenshot = await testPageScreenshot({
+      style,
       selector: '[data-dnb-test="global-status"] .dnb-global-status'
     })
     expect(screenshot).toMatchImageSnapshot()
   })
-  it.skip('have to match the info state', async () => {
+  it('have to match the close button in focus state', async () => {
     const screenshot = await testPageScreenshot({
+      style,
+      selector: '[data-dnb-test="global-status"] .dnb-global-status',
+      simulateSelector:
+        '[data-dnb-test="global-status"] .dnb-global-status__close-button',
+      simulate: 'focus'
+    })
+    expect(screenshot).toMatchImageSnapshot()
+  })
+  it('have to match the close button in hover state', async () => {
+    const screenshot = await testPageScreenshot({
+      style,
+      selector: '[data-dnb-test="global-status"] .dnb-global-status',
+      simulateSelector:
+        '[data-dnb-test="global-status"] .dnb-global-status__close-button',
+      simulate: 'hover'
+    })
+    expect(screenshot).toMatchImageSnapshot()
+  })
+  it('have to match the info state and custom content', async () => {
+    const screenshot = await testPageScreenshot({
+      style,
       selector: '[data-dnb-test="global-status-info"] .dnb-global-status'
     })
     expect(screenshot).toMatchImageSnapshot()
   })
-  it.skip('have to match custom content', async () => {
-    const screenshot = await testPageScreenshot({
-      selector: '[data-dnb-test="global-status-custom"]'
-    })
-    expect(screenshot).toMatchImageSnapshot()
-  })
+  // it.skip('have to match custom content', async () => {
+  //   const screenshot = await testPageScreenshot({
+  //     selector: '[data-dnb-test="global-status-custom"]'
+  //   })
+  //   expect(screenshot).toMatchImageSnapshot()
+  // })
 })